+++ alias_functions/gd.sh
@@ -5,12 +5,21 @@
 #
 # git diff
 
+. ~/.fancy-git/modules/settings-manager.sh
+git_use_clear=$(fancygit_config_get "git_use_clear" "true")
+
+clear_if_enabled () {
+    if [ "$git_use_clear" = "true" ]; then
+      clear
+    fi
+}
+
 # It's more like a return command. This is a Mac OS issue.
 fg_return() {
     exit
 }
 
-clear
+clear_if_enabled
 
 if [ "$1" = "" ]; then
     git diff
@@ -24,7 +33,7 @@ read -rp " Add this file to commit? [y/N]: " r
 
 if [ "$r" = "y" ]; then
     git add "$1"
-    clear
+    clear_if_enabled
     git status
     fg_return
 fi
@@ -35,5 +44,5 @@ if [ "$r" = "y" ]; then
     git checkout "$1"
 fi
 
-clear
+clear_if_enabled
 git status